requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
gunicorn
psycopg2-binary
stripe
celery
redis
django-celery-beat
python-dotenv
